Yjdh-267: check authentication every 5 minutes (#407)
* YJDH-267: check authentication every 5 minutes

* simplify withAuth logic

* run prettier
